\name{animal.noise}
\alias{animal.noise}
\title{Determine the noise of an animal}
\usage{
  animal.noise(animal)
}
\arguments{
  \item{animal}{name of animal}
}
\value{
  Animal's noise
}
\description{
  This function works out the appropriate noise
}
\examples{
  animal.noise("elephant") # Who cares
  animal.noise("dog") # Woof
}
\author{
  matt
}

