% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/the_data.R
\docType{data}
\name{breakfastDissimilarities}
\alias{breakfastDissimilarities}
\title{Breakfast Dissimilarities}
\format{A list containing four \code{\link[stats]{dist}} objects,
showing the perceived dissimilarities between 15 breafast foods

#'}
\source{
Roxygen James M. Lattin, J. Douglas Carroll, Paul E. Green (2003): Analyzing Multivariate Data, Thomson Brooks/Cole.
}
\usage{
breakfastDissimilarities
}
\description{
A dataset containing the proportion of people in a sample to associate various brands with various attributes
with different brands
#'
}
\keyword{datasets}

