% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/the_data.R
\docType{data}
\name{colaPerceptions}
\alias{colaPerceptions}
\title{Cola Perceptions}
\format{A matrix with 7 rows (6 brands and a 'none of these') and 9 columns (attributes)}
\usage{
colaPerceptions
}
\description{
A matrix displaying the proportion of people in a sample to associate various brands with various attributes
with different brands fo cola.
#'
}
\keyword{datasets}

