% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/the_data.R
\docType{data}
\name{csdPerceptions}
\alias{csdPerceptions}
\title{Carbonated Soft Drink Brand associations}
\format{A matrix with 8 rows (brands) and 11 columns (attributes)}
\usage{
csdPerceptions
}
\description{
A matrix displaying the proportion of people in a sample to associate various brands with various attributes
with different brands of carbonated soft drink (CSD).
#'
}
\keyword{datasets}

