% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/the_data.R
\docType{data}
\name{qColors}
\alias{qColors}
\title{Q's color palate (16 October 2014)}
\format{A vector containing RGB colors}
\usage{
qColors
}
\description{
The colors used in the data analysis package www.q-researchsoftware.com
#'
}
\keyword{datasets}

