% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animal.noise.R
\name{animal.noise}
\alias{animal.noise}
\title{Determine the noise of an animal}
\usage{
animal.noise(animal)
}
\arguments{
\item{animal}{name of animal}
}
\value{
Animal's noise
}
\description{
This function works out the appropriate noise
}
\examples{
animal.noise("elephant") # Silence
animal.noise("dog") # Woof
}
\author{
matt
}

